+++ paddler_bootstrap/src/agent_runner.rs
@@ -14,7 +14,7 @@ pub struct AgentRunnerParams {
     pub agent_name: Option<String>,
     pub cancellation_token: CancellationToken,
     pub management_address: String,
-    pub slots: i32,
+    pub slots: u16,
 }
 
 pub struct AgentRunner {
