+++ paddler_cluster/src/agent_config.rs
@@ -1,12 +1,12 @@
 #[derive(Clone, Debug)]
 pub struct AgentConfig {
     pub name: String,
-    pub slot_count: i32,
+    pub slot_count: u16,
 }
 
 impl AgentConfig {
     #[must_use]
-    pub fn single(slot_count: i32) -> Self {
+    pub fn single(slot_count: u16) -> Self {
         Self {
             name: "test-agent".to_owned(),
             slot_count,
@@ -14,7 +14,7 @@ impl AgentConfig {
     }
 
     #[must_use]
-    pub fn uniform(count: usize, slot_count: i32) -> Vec<Self> {
+    pub fn uniform(count: usize, slot_count: u16) -> Vec<Self> {
         (0..count)
             .map(|agent_index| Self {
                 name: format!("test-agent-{agent_index}"),
