+++ paddler_messaging/src/oversized_embedding_document_details.rs
@@ -4,7 +4,7 @@ use serde::Serialize;
 #[derive(Debug, Deserialize, Serialize)]
 #[serde(deny_unknown_fields)]
 pub struct OversizedEmbeddingDocumentDetails {
-    pub document_tokens: u32,
-    pub n_batch: u32,
+    pub document_tokens: usize,
+    pub n_batch: usize,
     pub source_document_id: String,
 }
